pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-agent-sdk"
version = "0.1.0"
description = "Generic framework for building autonomous AI agents with pluggable integrations (memory, governance, tasks). Includes YARNNN integration for governed long-term memory."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "YARNNN Contributors"}
]
keywords = ["ai", "agents", "claude", "autonomous", "memory", "governance", "yarnnn", "pluggable", "integrations", "framework"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "anthropic>=0.18.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
]
all = [
    "claude-agent-sdk[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/Kvkthecreator/claude-agentsdk-yarnn"
Documentation = "https://github.com/Kvkthecreator/claude-agentsdk-yarnn/tree/main/docs"
Repository = "https://github.com/Kvkthecreator/claude-agentsdk-yarnn"
"Bug Tracker" = "https://github.com/Kvkthecreator/claude-agentsdk-yarnn/issues"

[tool.setuptools]
packages = [
    "claude_agent_sdk",
    "claude_agent_sdk.archetypes",
    "claude_agent_sdk.integrations",
    "claude_agent_sdk.integrations.memory",
    "claude_agent_sdk.integrations.yarnnn",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
